libkvs/watch: split watch out to its own module
